Boobs update.
Added a different organ in replace to torso. Torsoboobs. wich allows you
to use the boobs variable. only humans have it so far, but if we go the
heads route, it won't matter. alternatively we could make Angel sprite
all the species :^ )
also fixed a bug with the subtle code that wasn't letting me compile,
with a badly named proc, hope I didn't broke anything.